test: Improve SSE coverage (#2956)
* Improve sse coverage

---------
